Move icon/aerosol/ to aerosol/icon/
Update absolute imports in test file, icon_physics.py, icon_terms.py,
parameters.py, and aerosol_radiation_test.py cross-reference.
