+++ jcm/physics/icon/icon_terms.py
@@ -170,7 +170,7 @@ class IconAerosol(IconTermBase):
     def __call__(self, state, diagnostics, forcing, terrain):
         """Compute aerosol optical properties."""
         data = self._build_data(diagnostics)
-        from jcm.physics.icon.aerosol.simple_aerosol import (
+        from jcm.physics.aerosol.icon.simple_aerosol import (
             get_simple_aerosol,
         )
         tend, data = get_simple_aerosol(
