+++ jcm/physics/icon/parameters.py
@@ -14,7 +14,7 @@ from jcm.physics.gravity_waves.icon import GravityWaveParameters
 from .radiation import RadiationParameters
 from .vertical_diffusion.vertical_diffusion_types import VDiffParameters
 from .surface import SurfaceParameters
-from .aerosol.aerosol_params import AerosolParameters
+from jcm.physics.aerosol.icon.aerosol_params import AerosolParameters
 
 @tree_math.struct
 class Parameters:
